tapai_extractor/requirements.txt
bitarray==2.8.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
google-ai-generativelanguage==0.2.0
google-api-core==2.11.1
google-auth==2.22.0
google-generativeai==0.1.0
googleapis-common-protos==1.60.0
grpcio==1.56.2
grpcio-status==1.56.2
idna==3.4
pdfreader==0.1.12
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.18.0
pypdf==3.12.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
urllib3==1.26.16
